Avoid setting the default runner
Setting the default runner to tox's default when `--current-env` is not
used can override another plugin like `tox-uv` which wants to change the
default runner. According to [this
comment](https://github.com/fedora-python/tox-current-env/pull/42#discussion_r1048595612)
the issue that led to the need to set the default runner has been
addressed and that appears to be confirmed by testing.